main: implement layer_surface::closed() event
Destroy the surface, and clear the ‘configured’ flag.

Note that we need to take care we don’t reference a destroyed output
object; if the compositor destroyed the output before calling the
closed() event, the ‘data’ argument will be an invalid pointer.

Since removing the output global _also_ destroys the surface, we can
handle this by looping all known output globals, and explicitly
destroy the surface if we find a match.

If we don’t find a match, that means the output has already been
destroyed, and we don’t have to do anything at all.